refactor: checkRadiobutton 컴포넌트를 지역 컴포넌트로 변경 - 한 군데에서만 사용되는 컴포넌트라 재사용성을 없애고 로컬 컴포넌트로 이동 - 응집도를 높임
